Cloduinary DeleteFiles + workspaces + tests

+++ app/Http/Controllers/Api/v1/MaterialController.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Http\Controllers\Api\v1;
 
+use App\Core\Resources\Storage\Storage;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\v1\Materials\CreateMaterialRequest;
 use App\Http\Requests\Api\v1\Materials\CreateTagRequest;
@@ -171,6 +172,8 @@ class MaterialController extends Controller
         }
 
         try {
+
+
             $data = removeNull([
                 'name' => $request->get('name')
             ]);
@@ -208,6 +211,14 @@ class MaterialController extends Controller
             ], 404);
         }
         try {
+            $result = Workspace::deleteFromStorage($workspace);
+
+            if (isset($result['error'])) {
+                return response()->json([
+                    'status' => 'error',
+                    'error' => $result['error']
+                ], 424);
+            }
 
             /**
              * WARNING: This will delete all the files assigned to this workspace
@@ -357,6 +368,7 @@ class MaterialController extends Controller
      * @apiResource App\Http\Resources\Api\Material\v1\MaterialResource
      * @apiResourceModel app\Models\Material
      * @response status=404 scenario="Material Not found"
+     * @response status=424 scenario="Override URL fail, we can not delete file from the source"
      */
     public function putEditMaterial(EditMaterialRequest $request, $materialId)
     {
@@ -375,6 +387,18 @@ class MaterialController extends Controller
                 'tags' => $request->get('tags') ? join(',', $request->get('tags')) : null,
                 'url' => $request->get('url')
             ]);
+
+            // Overriding URL => Delete old file
+            if ($request->get('url') && $request->get('url') !== $material->url) {
+                $result = Material::deleteFromStorage($material);
+
+                if (isset($result['error'])) {
+                    return response()->json([
+                        'status' => 'error',
+                        'error' => 'Error deleting file ' . $result['error']
+                    ], 424);
+                }
+            }
 
             Material::query()->find($material->id)->update($data);
             $updated = Material::query()->find($material->id);
@@ -398,6 +422,7 @@ class MaterialController extends Controller
      * @urlParam materialId integer Material Id
      * @response {"message": "successfully"}
      * @response status=404 scenario="Material Not found"
+     * @response status=424 scenario="Delete material fail: we can not delete URL from the source"
      */
     public function deleteMaterial($materialId): JsonResponse
     {
@@ -411,6 +436,18 @@ class MaterialController extends Controller
                     'error' => 'Material not found'
                 ], 404);
             }
+
+
+            // Deleting Material URL
+            $result = Material::deleteFromStorage($material);
+
+            if (isset($result['error'])) {
+                return response()->json([
+                    'status' => 'error',
+                    'error' => 'Error deleting file ' . $result['error']
+                ], 424);
+            }
+
 
             $material->delete();
